# Improve performance of NamespaceFilter when there are thousands of entries