<!-- This template is for Devs to give QA details before moving the issue To-Test -->
### Summary
Fixes #7775
<!-- Define findings related to the feature or bug issue. -->

### Occurred changes and/or fixed issues
- When there are thousands of entries certain actions (drop down opened, selection changed, etc) take a long time to complete (upwards of 5 seconds)
- This is caused by churn of the filtered and options computed properties causing multiple renders for each action.
- To break this multiple-render per cycle behaviour detach `filtered` from the value used in `v-for`.
- This knocks off 2/3 of the time taken for the actions. 

### Technical notes summary
- This is a lo-bar fix to address in the 2.7.next1 timeframe. In the future we should consider
  - only rendering what's in view 
  - not including whole model in options

### Areas or cases that should be tested
<!-- Areas that should be tested can include Airgap checks, Rancher upgrades, K8s upgrade, etc. -->
<!-- Which browser did you use for local testing? The reviewer should test with a different browser. -->
<!-- Add missing steps or rewrite them if have been missed or to complement existing information. This should define a clear way to reproduce it and not an approximation. -->

### Areas which could experience regressions
- Cluster Explorer view - Namespace Filter (opening, listing projects/namespaces, selecting projects/namespaces)

### Screenshot/Video
Process - Open up NS filter and change from user to system namespaces
**Before Change**
Open NS filter - ~3 seconds
Change selection - ~3 seconds
![image](https://user-images.githubusercontent.com/18697775/209125396-b88d5587-db51-4d34-8110-a172a072cf12.png)


**After Change**
Open NS filter - ~0.7 seconds
Change selection - ~1 second
![image](https://user-images.githubusercontent.com/18697775/209125046-d91e5570-bf5e-4c11-99d5-1f838259d82a.png)