@richard-cox 2022-12-23
I've just noticed we bring in a virtual scroll dep with https://github.com/rancher/dashboard/pull/7511. Next step would be to investigate using that